Correctly ignore C-style cast in CUDA header, short int->int to avoid information loss in type conversion

+++ splines/gpuSplineUtils.cu
@@ -211,7 +211,7 @@ __host__ void SMonolithGPU::CopyToGPU_SplineMonolith(
                     #endif
                             int n_params, 
                             unsigned int n_splines,
-                            short int spline_size,
+                            int spline_size,
                             unsigned int total_nknots,
                             unsigned int n_tf1) {
   if (n_params != _N_SPLINES_) {
